refactor(CrewRepository): 패키치 위치 이동
