+++ src/main/java/controller/AttendanceController.java
@@ -4,7 +4,6 @@ import constant.CampusConstant;
 import domain.AttendanceStatus;
 import domain.AttendanceStatusStatistics;
 import domain.Crew;
-import domain.CrewRepository;
 import domain.Manage;
 import dto.AttendanceModifyRequest;
 import dto.AttendanceRecord;
@@ -18,6 +17,7 @@ import dto.OptionRequest;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
+import repository.CrewRepository;
 import util.DateTimeUtil;
 import view.InputView;
 import view.OutputView;
