+++ src/test/java/domain/CrewRepositoryTest.java
@@ -1,14 +1,10 @@
 package domain;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import repository.CrewRepository;
 
 public class CrewRepositoryTest {
     @BeforeEach
