+++ src/test/java/domain/CrewTest.java
@@ -9,6 +9,7 @@ import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import repository.CrewRepository;
 
 public class CrewTest {
